Adding testing for inheritance
