+++ contracts/interfaces/IProtected.sol
