+++ contracts/manager/Manager.sol
@@ -47,7 +47,7 @@ contract Manager is IManager, Actor, Context, Versioned, ERC721Holder, UUPSUpgra
   error QuorumCannotBeZero();
   error QuorumCannotBeGreaterThanSentinels();
   error InheritanceNotConfigured();
-  error NotExpiredYet();
+  error StillAlive();
   error InconsistentRecipient();
   error NotASentinel();
   error RequestAlreadyApproved();
@@ -56,6 +56,9 @@ contract Manager is IManager, Actor, Context, Versioned, ERC721Holder, UUPSUpgra
   error CannotBeYourself();
   error NotTheFirstProtector();
   error FirstProtectorNotFound();
+
+  bool public constant IS_MANAGER = true;
+  bool public constant IS_NOT_MANAGER = false;
 
   IAccountGuardian public guardian;
   SignatureValidator public signatureValidator;
@@ -162,6 +165,10 @@ contract Manager is IManager, Actor, Context, Versioned, ERC721Holder, UUPSUpgra
     return _listActiveActors(PROTECTOR);
   }
 
+  function hasProtectors() public view override returns (bool) {
+    return _actorLength(PROTECTOR) > 0;
+  }
+
   // @dev see {IManager-setProtector}
   function setProtector(
     address protector_,
@@ -170,8 +177,17 @@ contract Manager is IManager, Actor, Context, Versioned, ERC721Holder, UUPSUpgra
     uint256 validFor,
     bytes calldata signature
   ) external virtual override onlyTokenOwner {
-    _setSignedActor("PROTECTOR", protector_, PROTECTOR, status, timestamp, validFor, signature, true);
+    _setSignedActor("PROTECTOR", protector_, PROTECTOR, status, timestamp, validFor, signature, IS_MANAGER);
     emit ProtectorUpdated(_msgSender(), protector_, status);
+    if (status) {
+      if (countActiveProtectors() == 1) {
+        vault.emitLockedEvent(tokenId(), true);
+      }
+    } else {
+      if (countActiveProtectors() == 0) {
+        vault.emitLockedEvent(tokenId(), false);
+      }
+    }
   }
 
   // @dev see {IManager-getProtectors}
@@ -188,7 +204,7 @@ contract Manager is IManager, Actor, Context, Versioned, ERC721Holder, UUPSUpgra
     uint256 validFor,
     bytes calldata signature
   ) external override onlyTokenOwner {
-    _setSignedActor("SAFE_RECIPIENT", recipient, SAFE_RECIPIENT, status, timestamp, validFor, signature, false);
+    _setSignedActor("SAFE_RECIPIENT", recipient, SAFE_RECIPIENT, status, timestamp, validFor, signature, IS_NOT_MANAGER);
     emit SafeRecipientUpdated(_msgSender(), recipient, status);
   }
 
@@ -202,7 +218,7 @@ contract Manager is IManager, Actor, Context, Versioned, ERC721Holder, UUPSUpgra
     return getActors(SAFE_RECIPIENT);
   }
 
-  // beneficiaries
+  // sentinels and beneficiaries
   // @dev see {IManager-setSentinel}
   function setSentinel(
     address sentinel,
@@ -211,7 +227,7 @@ contract Manager is IManager, Actor, Context, Versioned, ERC721Holder, UUPSUpgra
     uint256 validFor,
     bytes calldata signature
   ) external override onlyTokenOwner {
-    _setSignedActor("SENTINEL", sentinel, SENTINEL, status, timestamp, validFor, signature, false);
+    _setSignedActor("SENTINEL", sentinel, SENTINEL, status, timestamp, validFor, signature, IS_NOT_MANAGER);
     emit SentinelUpdated(_msgSender(), sentinel, status);
   }
 
@@ -224,11 +240,16 @@ contract Manager is IManager, Actor, Context, Versioned, ERC721Holder, UUPSUpgra
     // solhint-disable-next-line not-rely-on-time
     _inheritanceConf = InheritanceConf(quorum, proofOfLifeDurationInDays, block.timestamp);
     delete _inheritanceRequest;
+    emit InheritanceConfigured(_msgSender(), quorum, proofOfLifeDurationInDays);
   }
 
-  // @dev see {IManager-getSentinels}
-  function getSentinels() external view returns (address[] memory, InheritanceConf memory) {
-    return (getActors(SENTINEL), _inheritanceConf);
+  // @dev see {IManager-getSentinelsAndInheritanceData}
+  function getSentinelsAndInheritanceData()
+    external
+    view
+    returns (address[] memory, InheritanceConf memory, InheritanceRequest memory)
+  {
+    return (getActors(SENTINEL), _inheritanceConf, _inheritanceRequest);
   }
 
   // @dev see {IManager-proofOfLife}
@@ -237,6 +258,7 @@ contract Manager is IManager, Actor, Context, Versioned, ERC721Holder, UUPSUpgra
     // solhint-disable-next-line not-rely-on-time
     _inheritanceConf.lastProofOfLife = block.timestamp;
     delete _inheritanceRequest;
+    emit ProofOfLife(_msgSender());
   }
 
   // @dev see {IManager-requestTransfer}
@@ -246,37 +268,47 @@ contract Manager is IManager, Actor, Context, Versioned, ERC721Holder, UUPSUpgra
     uint256 i = _findActorIndex(_msgSender(), SENTINEL);
     if (i == MAX_ACTORS) revert NotASentinel();
     if (
-      _inheritanceConf.lastProofOfLife + (_inheritanceConf.proofOfLifeDurationInDays * 1 hours) >
+      _inheritanceConf.lastProofOfLife + (_inheritanceConf.proofOfLifeDurationInDays * 1 days) >
       // solhint-disable-next-line not-rely-on-time
       block.timestamp
-    ) revert NotExpiredYet();
+    ) revert StillAlive();
     // the following prevents hostile beneficiaries from blocking the process not allowing them to reset the recipient
-    if (_hasApproved()) revert RequestAlreadyApproved();
-    // a sentinel is proposing a new recipient
-    if (_inheritanceRequest.recipient != beneficiary) {
-      // solhint-disable-next-line not-rely-on-time
+    for (i = 0; i < _inheritanceRequest.approvers.length; i++) {
+      if (_msgSender() == _inheritanceRequest.approvers[i]) {
+        revert RequestAlreadyApproved();
+      }
+    }
+    if (_inheritanceRequest.beneficiary != beneficiary) {
+      // a sentinel can propose a new beneficiary only after the first request expires
       if (block.timestamp - _inheritanceRequest.startedAt > 30 days) {
-        // reset the request
         delete _inheritanceRequest;
       } else revert InconsistentRecipient();
     }
-    if (_inheritanceRequest.recipient == address(0)) {
-      _inheritanceRequest.recipient = beneficiary;
+    if (_inheritanceRequest.beneficiary == address(0)) {
+      _inheritanceRequest.beneficiary = beneficiary;
       // solhint-disable-next-line not-rely-on-time
       _inheritanceRequest.startedAt = block.timestamp;
       _inheritanceRequest.approvers.push(_msgSender());
+      emit TransferRequested(_msgSender(), beneficiary);
     } else {
       _inheritanceRequest.approvers.push(_msgSender());
+      emit TransferRequestApproved(_msgSender());
     }
   }
 
-  // TODO add a deadline after a while a new beneficiary can be set
   // @dev see {IManager-inherit}
   function inherit() external {
-    if (_inheritanceRequest.recipient == _msgSender() && _inheritanceRequest.approvers.length >= _inheritanceConf.quorum) {
-      vault.managedTransfer(tokenId(), _msgSender());
-      emit Inherited(tokenAddress(), tokenId(), owner(), _msgSender());
+    // we set an expiration time in case the beneficiary cannot inherit
+    // so the sentinels can propose a new beneficiary
+    if (block.timestamp - _inheritanceRequest.startedAt > 60 days) {
       delete _inheritanceRequest;
+    }
+    if (_inheritanceRequest.beneficiary == _msgSender() && _inheritanceRequest.approvers.length >= _inheritanceConf.quorum) {
+      delete _inheritanceConf;
+      delete _inheritanceRequest;
+      _resetActors();
+      vault.managedTransfer(tokenId(), _msgSender());
+      emit InheritedBy(_msgSender());
     } else revert Unauthorized();
   }
 
@@ -346,16 +378,6 @@ contract Manager is IManager, Actor, Context, Versioned, ERC721Holder, UUPSUpgra
       if (timestamp != 0 && actorIsProtector && isAProtector(actor)) revert ProtectorAlreadySetByYou();
       _addActor(actor, role_);
     }
-  }
-
-  // @dev Returns true if the sentinels have approved a request.
-  function _hasApproved() internal view returns (bool) {
-    for (uint256 i = 0; i < _inheritanceRequest.approvers.length; i++) {
-      if (_msgSender() == _inheritanceRequest.approvers[i]) {
-        return true;
-      }
-    }
-    return false;
   }
 
   // @dev This empty reserved space is put in place to allow future versions to add new
