+++ export/ABIs.json
@@ -1,5 +1,5 @@
 {
-  "when": "2023-11-25T18:59:03.871Z",
+  "when": "2023-11-25T22:55:32.992Z",
   "contracts": {
     "ERC6551Registry": [
       {
@@ -203,11 +203,6 @@
       },
       {
         "inputs": [],
-        "name": "NotExpiredYet",
-        "type": "error"
-      },
-      {
-        "inputs": [],
         "name": "NotPermittedWhenProtectorsAreActive",
         "type": "error"
       },
@@ -259,6 +254,11 @@
       {
         "inputs": [],
         "name": "SignatureAlreadyUsed",
+        "type": "error"
+      },
+      {
+        "inputs": [],
+        "name": "StillAlive",
         "type": "error"
       },
       {
@@ -334,29 +334,49 @@
           {
             "indexed": true,
             "internalType": "address",
-            "name": "protected",
+            "name": "owner",
             "type": "address"
           },
           {
             "indexed": false,
             "internalType": "uint256",
-            "name": "tokenId",
+            "name": "quorum",
             "type": "uint256"
           },
           {
-            "indexed": true,
-            "internalType": "address",
-            "name": "from",
-            "type": "address"
-          },
+            "indexed": false,
+            "internalType": "uint256",
+            "name": "proofOfLifeDurationInDays",
+            "type": "uint256"
+          }
+        ],
+        "name": "InheritanceConfigured",
+        "type": "event"
+      },
+      {
+        "anonymous": false,
+        "inputs": [
           {
             "indexed": true,
             "internalType": "address",
-            "name": "to",
+            "name": "beneficiary",
             "type": "address"
           }
         ],
-        "name": "Inherited",
+        "name": "InheritedBy",
+        "type": "event"
+      },
+      {
+        "anonymous": false,
+        "inputs": [
+          {
+            "indexed": true,
+            "internalType": "address",
+            "name": "owner",
+            "type": "address"
+          }
+        ],
+        "name": "ProofOfLife",
         "type": "event"
       },
       {
@@ -440,12 +460,70 @@
           {
             "indexed": true,
             "internalType": "address",
+            "name": "sentinel",
+            "type": "address"
+          }
+        ],
+        "name": "TransferRequestApproved",
+        "type": "event"
+      },
+      {
+        "anonymous": false,
+        "inputs": [
+          {
+            "indexed": true,
+            "internalType": "address",
+            "name": "sentinel",
+            "type": "address"
+          },
+          {
+            "indexed": true,
+            "internalType": "address",
+            "name": "beneficiary",
+            "type": "address"
+          }
+        ],
+        "name": "TransferRequested",
+        "type": "event"
+      },
+      {
+        "anonymous": false,
+        "inputs": [
+          {
+            "indexed": true,
+            "internalType": "address",
             "name": "implementation",
             "type": "address"
           }
         ],
         "name": "Upgraded",
         "type": "event"
+      },
+      {
+        "inputs": [],
+        "name": "IS_MANAGER",
+        "outputs": [
+          {
+            "internalType": "bool",
+            "name": "",
+            "type": "bool"
+          }
+        ],
+        "stateMutability": "view",
+        "type": "function"
+      },
+      {
+        "inputs": [],
+        "name": "IS_NOT_MANAGER",
+        "outputs": [
+          {
+            "internalType": "bool",
+            "name": "",
+            "type": "bool"
+          }
+        ],
+        "stateMutability": "view",
+        "type": "function"
       },
       {
         "inputs": [],
@@ -596,7 +674,7 @@
       },
       {
         "inputs": [],
-        "name": "getSentinels",
+        "name": "getSentinelsAndInheritanceData",
         "outputs": [
           {
             "internalType": "address[]",
@@ -624,6 +702,28 @@
             "internalType": "struct IManager.InheritanceConf",
             "name": "",
             "type": "tuple"
+          },
+          {
+            "components": [
+              {
+                "internalType": "address",
+                "name": "beneficiary",
+                "type": "address"
+              },
+              {
+                "internalType": "uint256",
+                "name": "startedAt",
+                "type": "uint256"
+              },
+              {
+                "internalType": "address[]",
+                "name": "approvers",
+                "type": "address[]"
+              }
+            ],
+            "internalType": "struct IManager.InheritanceRequest",
+            "name": "",
+            "type": "tuple"
           }
         ],
         "stateMutability": "view",
@@ -637,6 +737,19 @@
             "internalType": "contract IAccountGuardian",
             "name": "",
             "type": "address"
+          }
+        ],
+        "stateMutability": "view",
+        "type": "function"
+      },
+      {
+        "inputs": [],
+        "name": "hasProtectors",
+        "outputs": [
+          {
+            "internalType": "bool",
+            "name": "",
+            "type": "bool"
           }
         ],
         "stateMutability": "view",
@@ -1225,6 +1338,38 @@
         "anonymous": false,
         "inputs": [
           {
+            "indexed": false,
+            "internalType": "bool",
+            "name": "locked",
+            "type": "bool"
+          }
+        ],
+        "name": "DefaultLocked",
+        "type": "event"
+      },
+      {
+        "anonymous": false,
+        "inputs": [
+          {
+            "indexed": true,
+            "internalType": "uint256",
+            "name": "tokenId",
+            "type": "uint256"
+          },
+          {
+            "indexed": false,
+            "internalType": "bool",
+            "name": "locked",
+            "type": "bool"
+          }
+        ],
+        "name": "Locked",
+        "type": "event"
+      },
+      {
+        "anonymous": false,
+        "inputs": [
+          {
             "indexed": true,
             "internalType": "address",
             "name": "previousOwner",
@@ -1395,6 +1540,37 @@
       },
       {
         "inputs": [],
+        "name": "defaultLocked",
+        "outputs": [
+          {
+            "internalType": "bool",
+            "name": "",
+            "type": "bool"
+          }
+        ],
+        "stateMutability": "pure",
+        "type": "function"
+      },
+      {
+        "inputs": [
+          {
+            "internalType": "uint256",
+            "name": "tokenId",
+            "type": "uint256"
+          },
+          {
+            "internalType": "bool",
+            "name": "locked_",
+            "type": "bool"
+          }
+        ],
+        "name": "emitLockedEvent",
+        "outputs": [],
+        "stateMutability": "nonpayable",
+        "type": "function"
+      },
+      {
+        "inputs": [],
         "name": "factoryAddress",
         "outputs": [
           {
@@ -1468,6 +1644,25 @@
           }
         ],
         "name": "isTransferable",
+        "outputs": [
+          {
+            "internalType": "bool",
+            "name": "",
+            "type": "bool"
+          }
+        ],
+        "stateMutability": "view",
+        "type": "function"
+      },
+      {
+        "inputs": [
+          {
+            "internalType": "uint256",
+            "name": "tokenId",
+            "type": "uint256"
+          }
+        ],
+        "name": "locked",
         "outputs": [
           {
             "internalType": "bool",
