+++ tmp/coverage.report
@@ -13,13 +13,14 @@ Instrumenting for coverage...
 > CrunaFlexiVault.sol
 > factory/IVaultFactory.sol
 > factory/VaultFactory.sol
+> interfaces/IERC6454.sol
+> interfaces/IERC6982.sol
+> interfaces/IProtected.sol
 > manager/Actor.sol
 > manager/Guardian.sol
 > manager/IManager.sol
 > manager/Manager.sol
 > manager/ManagerProxy.sol
-> protected/IERC6454.sol
-> protected/IProtected.sol
 > protected/ProtectedNFT.sol
 > utils/SignatureValidator.sol
 > utils/UUPSUpgradableTemplate.sol
@@ -44,29 +45,33 @@ Network Info
 
 
 
-  Integration
+  Testing contract deployments
     ✔ should deploy everything as expected
-    ✔ should get the token parameters from the manager (55ms)
-    Protectors management
-      ✔ should add the first protector (62ms)
-      ✔ should throw is wrong data for first protector (84ms)
-      ✔ should add many protectors (127ms)
-      ✔ should add a protector and transfer a vault (95ms)
-    Safe recipients
-      ✔ should set up safe recipients (191ms)
-    Sentinels and beneficiary
-      ✔ should set up sentinel (199ms)
+    ✔ should get the token parameters from the manager (70ms)
+
+  Integration
+    ✔ should add the first protector (70ms)
+    ✔ should throw is wrong data for first protector (101ms)
+    ✔ should add many protectors (150ms)
+    ✔ should add a protector and transfer a vault (118ms)
+
+  Protectors
+    ✔ should set up safe recipients (206ms)
+
+  Integration
+    ✔ should set up sentinel (229ms)
+    ✔ should set up 5 sentinels and an inheritance with a quorum 3 (312ms)
 
   SignatureValidator
     ✔ should recover the signer of a recoverSigner
 
   Factory
-    ✔ should get the precalculated address of the manager
-    ✔ should allow bob and alice to purchase some vaults (109ms)
-    ✔ should allow bob and alice to purchase some vaults with a promoCode (99ms)
+    ✔ should get the precalculated address of the manager (41ms)
+    ✔ should allow bob and alice to purchase some vaults (128ms)
+    ✔ should allow bob and alice to purchase some vaults with a promoCode (157ms)
 
 
-  12 passing (3s)
+  13 passing (4s)
 
 -----------------------------|----------|----------|----------|----------|----------------|
 File                         |  % Stmts | % Branch |  % Funcs |  % Lines |Uncovered Lines |
@@ -76,22 +81,24 @@ File                         |  % Stmts | % Branch |  % Funcs |  % Lines |Uncove
  contracts/factory/          |    57.14 |     37.5 |    81.82 |    58.18 |                |
   IVaultFactory.sol          |      100 |      100 |      100 |      100 |                |
   VaultFactory.sol           |    57.14 |     37.5 |    81.82 |    58.18 |... 126,130,143 |
- contracts/manager/          |    56.99 |    33.33 |    58.97 |    55.66 |                |
-  Actor.sol                  |    94.74 |       60 |     87.5 |       95 |             49 |
+ contracts/interfaces/       |      100 |      100 |      100 |      100 |                |
+  IERC6454.sol               |      100 |      100 |      100 |      100 |                |
+  IERC6982.sol               |      100 |      100 |      100 |      100 |                |
+  IProtected.sol             |      100 |      100 |      100 |      100 |                |
+ contracts/manager/          |    81.37 |    58.33 |     72.5 |    83.47 |                |
+  Actor.sol                  |    94.74 |       60 |    88.89 |    95.65 |             49 |
   Guardian.sol               |      100 |      100 |        0 |      100 |                |
   IManager.sol               |      100 |      100 |      100 |      100 |                |
-  Manager.sol                |     47.3 |       30 |    51.72 |    46.51 |... 354,355,358 |
+  Manager.sol                |    78.31 |    58.14 |    68.97 |    80.61 |... 218,278,304 |
   ManagerProxy.sol           |      100 |      100 |      100 |      100 |                |
- contracts/protected/        |       76 |       35 |       70 |    74.29 |                |
-  IERC6454.sol               |      100 |      100 |      100 |      100 |                |
-  IProtected.sol             |      100 |      100 |      100 |      100 |                |
-  ProtectedNFT.sol           |       76 |       35 |       70 |    74.29 |... 121,122,138 |
+ contracts/protected/        |    86.21 |    42.86 |    76.92 |    89.74 |                |
+  ProtectedNFT.sol           |    86.21 |    42.86 |    76.92 |    89.74 | 81,140,165,174 |
  contracts/utils/            |    92.31 |    66.67 |    66.67 |    81.82 |                |
   SignatureValidator.sol     |      100 |     87.5 |      100 |     87.5 |             31 |
   UUPSUpgradableTemplate.sol |      100 |       25 |       50 |      100 |                |
   Versioned.sol              |        0 |      100 |        0 |        0 |              9 |
 -----------------------------|----------|----------|----------|----------|----------------|
-All files                    |    62.36 |       37 |    65.28 |    61.21 |                |
+All files                    |    76.96 |       50 |    73.68 |    78.11 |                |
 -----------------------------|----------|----------|----------|----------|----------------|
 
 > Istanbul reports written to ./coverage/ and ./coverage.json
